% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setEnsemblePaths.R
\name{build_pathset}
\alias{build_pathset}
\title{Construct paths to a series of similarly-named files}
\usage{
build_pathset(n, glue_str = "./file_{n}.nc", ...)
}
\arguments{
\item{n}{vector of replicate ids.
If given as a single numeric value, it will be expanded to \code{1:n}}

\item{glue_str}{Specification of how to interpolate variables into the
paths, as a \code{glue::glue()} input (see examples)}

\item{...}{other variables to be interpolated into the path, each with length
either 1 or equal to \code{n}.}
}
\value{
list of paths the same length as \code{n}, with names set as \verb{path<n>}
}
\description{
Given a template string and a set of variables that includes a replicate ID,
creates a list of paths matching the template and returns them in a list
ready to insert into a PEcAn \code{settings} object.
}
\details{
The motivating use case is populating the ensemble files (\code{met$path},
\code{poolinitcond$path}, etc) inside \verb{settings$run$site.<x>$inputs}
when creating multi-site run configurations. It can also be helpful for
e.g. series of yearly files, or more generally for constructing any set of
paths that both (1) have names differing by one variable, and (2) should be
inserted into a PEcAn settings object as one block of \verb{<path>} entries:

\if{html}{\out{<div class="sourceCode">}}\preformatted{<path>
  <path1>path/to/siteA/file.1.ext</path1>
  <path2>path/to/siteA/file.2.ext</path2>
  [...]
  <path[n]>path/to/siteA/file.[n].ext</path[n]>
</path>
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
build_pathset(3, "IC/foo/IC_foo_{n}.nc")
build_pathset(n = 1:3, id = "2ce800f4", yr = 2003, "ERA5/{id}/{yr}_{n}.clim")
}
}
\keyword{internal}
