% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SafeList.R
\name{$.SafeList}
\alias{$.SafeList}
\title{Extract SafeList component by name}
\usage{
\method{$}{SafeList}(x, name)
}
\arguments{
\item{x}{the SafeList object}

\item{name}{the name of the component}
}
\value{
The specified component
}
\description{
Extract SafeList component by name
}
\details{
Overrides \verb{$.list}, and works just like it except forces exact match
(i.e., makes \code{x$name} behave exactly like \code{x[[name, exact=T]]})
}
\author{
Ryan Kelly
}
