% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.all.settings.R
\name{check.bety.version}
\alias{check.bety.version}
\title{Check BETY Version}
\usage{
check.bety.version(dbcon)
}
\arguments{
\item{dbcon}{database connection object}
}
\description{
check to make sure BETY is up to date
}
