% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.all.settings.R
\name{check.database}
\alias{check.database}
\title{Check Database}
\usage{
check.database(database)
}
\arguments{
\item{database}{settings list to check.
You'll probably use \code{settings$database}}
}
\description{
Check Database
}
