% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.all.settings.R
\name{check.database.settings}
\alias{check.database.settings}
\title{Check Database Settings}
\usage{
check.database.settings(settings)
}
\arguments{
\item{settings}{settings file}
}
\description{
Check Database Settings
}
