% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.all.settings.R
\name{check.ensemble.settings}
\alias{check.ensemble.settings}
\title{Check ensemble Settings}
\usage{
check.ensemble.settings(settings)
}
\arguments{
\item{settings}{settings file}
}
\description{
Check ensemble Settings
}
