% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.inputs.R
\name{check.inputs}
\alias{check.inputs}
\title{Update Settings}
\usage{
check.inputs(settings)
}
\arguments{
\item{settings}{settings list}
}
\value{
updated settings list
}
\description{
check to see if inputs are specified - this should be part of the model code
}

