% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.all.settings.R
\name{check.model.settings}
\alias{check.model.settings}
\title{Check Model Settings}
\usage{
check.model.settings(settings, dbcon = NULL)
}
\arguments{
\item{settings}{settings file}

\item{dbcon}{database connection.}
}
\description{
Check Model Settings
}
