% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.all.settings.R
\name{check.run.settings}
\alias{check.run.settings}
\title{Check Run Settings}
\usage{
check.run.settings(settings, dbcon = NULL)
}
\arguments{
\item{settings}{settings file}

\item{dbcon}{database connection.}
}
\description{
Check Run Settings
}
