% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.all.settings.R
\name{check.workflow.settings}
\alias{check.workflow.settings}
\title{Check Workflow Settings}
\usage{
check.workflow.settings(settings, dbcon = NULL)
}
\arguments{
\item{settings}{settings file}

\item{dbcon}{database connection}
}
\description{
Check Workflow Settings
}
