% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMultisiteMultiSettings.r
\name{getRunSettings}
\alias{getRunSettings}
\title{Build run MultiSettings for a single site id}
\usage{
getRunSettings(templateSettings, siteId)
}
\arguments{
\item{templateSettings}{A \code{\link{Settings}} object that will be
the template for the resulting MultiSettings.}

\item{siteId}{site to process. See \code{createMultiSiteSettings}}
}
\description{
Processes one site from the \code{siteIds} argument of \code{createMultiSiteSettings}.
You probably don't need to call it directly.
}
