% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{inlist}
\alias{inlist}
\alias{inrange}
\title{Option checkers}
\usage{
inlist(...)

inrange(min = -Inf, max = Inf)
}
\arguments{
\item{...}{comma-separated list of allowed values.}

\item{min}{minimum value (for numeric options)}

\item{max}{maximum value (for numeric options)}
}
\description{
These functions return a function that is used by the options manager internally
to check whether an option set by the user is allowed.
}
\seealso{
\code{\link{options_manager}} for examples.
}
