% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
\name{is_setting}
\alias{is_setting}
\title{Find out if we're setting or getting}
\usage{
is_setting(...)
}
\arguments{
\item{...}{\code{[key]=[value]} pairs of options}
}
\value{
\code{logical}, \code{TRUE} if \code{...} represents set-options, \code{FALSE} if
 \code{...} represents get-options. An error is thrown if it cannot be determined.
}
\description{
Utility function checking if we're setting or getting.
}
\seealso{
\code{\link{stop_if_reserved}}
}

