% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listToXml.R
\name{listToXml}
\alias{listToXml}
\title{A generic function to convert list to XML}
\usage{
listToXml(x, ...)
}
\arguments{
\item{x}{list to be converted}

\item{...}{arguments passed to methods}
}
\description{
A generic function to convert list to XML
}
