% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.global.settings.R
\name{parse.global.settings}
\alias{parse.global.settings}
\title{Parse global settings into a list of pecan Settings objects}
\usage{
parse.global.settings(global.settings)
}
\arguments{
\item{global.settings}{The global settings object, likely returned by read.settings()}
}
\value{
A list of settings objects
}
\description{
Currently, used to convert a multi-site settings object into a list of settings objects, one
per site. Could be extended to handle multi-model comparisons or other applications
}
\author{
Ryan Kelly
}

