% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.settings.R
\name{prepare.settings}
\alias{prepare.settings}
\title{Prepare Settings}
\usage{
prepare.settings(settings, force = FALSE)
}
\arguments{
\item{settings}{settings list}

\item{force}{Whether to force the function to run even if it determines it has been run on 
these settings already.}
}
\description{
Update, set defaults for, and otherwise prepare a PEcAn Settings object
}
\details{
Performs various checks, fixes deprecated contructs, and assigns missing values where possible.
}
\author{
Ryan Kelly

Betsy Cowdery
}
