% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSettings.R
\name{printAll}
\alias{printAll}
\title{generic function for printing contents of objects.}
\usage{
printAll(x)
}
\arguments{
\item{x}{object to be printed.}
}
\description{
generic function for printing contents of objects.
}
