% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
\name{reset}
\alias{reset}
\title{Reset options to default values}
\usage{
reset(options)
}
\arguments{
\item{options}{An option manager, as returned by \code{\link{options_manager}} or \code{\link{clone_and_merge}}}
}
\value{
The list of reset options, invisibly.
}
\description{
Reset options to default values
}
\seealso{
\code{\link{defaults}}
}

