% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{reset_options}
\alias{reset_options}
\title{Reset general options in 'options' to factory defaults.}
\usage{
reset_options()
}
\description{
Reset general options in 'options' to factory defaults.
}
\seealso{
\code{\link{reset_par}}
}

