% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SafeList.R
\name{SafeList}
\alias{SafeList}
\alias{as.SafeList}
\alias{is.SafeList}
\title{Constrct SafeList Object}
\usage{
SafeList(...)

as.SafeList(x)

is.SafeList(x)
}
\arguments{
\item{...}{A list to upgrade to SafeList, or elements to be added to a new SafeList}

\item{x}{list to coerce}

\item{x}{list object to be tested}
}
\value{
The resulting SafeList

a SafeList version of x

logical
}
\description{
Create a SafeList object
}
\details{
SafeList is a wrapper class for the normal R list. It should behave identically, except for
the $ operator being overridden to require exact matches.

The constructor works identical to list(...) unless:

1) The only argument is a list, in which case the result is the same list, with its class
attribute updated to include 'SafeList', or
2) The only argument is a SafeList, in which case that argument is returned unchanged
}
\section{Functions}{
\itemize{
\item \code{as.SafeList}: Coerce an object to SafeList.

\item \code{is.SafeList}: Test if object is already a SafeList.
}}

\author{
Ryan Kelly
}
