% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMultisiteMultiSettings.r
\name{setDates}
\alias{setDates}
\title{Set the Dates of PEcAn Settings}
\usage{
setDates(settings, startDate, endDate)
}
\arguments{
\item{settings}{A \code{\link{Settings}} object}

\item{startDate, endDate}{The desired start and end dates}
}
\value{
The original \code{Settings} object with updated dates
}
\description{
Sets the run, ensemble, and sensitivity analysis dates of PEcAn Settings
}
\details{
Sets the start/end dates in \code{settings$run} to the specified dates, and sets the corresponding 
years for \code{settings$ensemble} and \code{settings$sensitivity.analysis}. Either date can be 
omitted to leave it unchanged.
}
\author{
Ryan Kelly
}
