% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMultisiteMultiSettings.r
\name{setOutDir}
\alias{setOutDir}
\title{Set the Ouptu Directories of PEcAn Settings}
\usage{
setOutDir(settings, outDir)
}
\arguments{
\item{settings}{A \code{\link{Settings}} object}

\item{outDir}{The desired output directory}
}
\value{
The original \code{Settings} object with updated output directories
}
\description{
Sets the main output directory and nulls out the others
}
\details{
Sets the main output directory (\code{settings$outdir}) to \code{outDir}, and sets numerous others
(\code{settings$modeloutdir}, \code{settings$host$rundir}, \code{settings$host$outdir}, 
\code{settings$host$modeloutdir}) to NULL so they will revert to defaults when 
\code{\link{check.settings}} is run.
}
\author{
Ryan Kelly
}
