% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSettings.R
\name{settingNames}
\alias{settingNames}
\title{function that can retrieve or update the names of multi-settings.}
\usage{
settingNames(multiSettings, settingNames)
}
\description{
function that can retrieve or update the names of multi-settings.
}
