% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/settings-package.R
\docType{package}
\name{settings}
\alias{package-settings}
\alias{settings}
\alias{settings-package}
\title{Convenient options settings management for R}
\description{
Convenient options settings management for R
}
\details{
Reset graphical options of \code{par} or \code{options} to their factory settings
using \code{\link{reset_par}} or \code{\link{reset_par}}.

Create your own option settings manager with \code{\link{options_manager}} for projects
with many options. Clone and merge an options manager for easy hierarchical options management
using \code{\link{clone_and_merge}}.

See the vignette for examples, type

\code{vignette("settings", package = "settings")}

at the command-line.
}

