% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stop_if_reserved}
\alias{stop_if_reserved}
\title{Check for reserved option names.}
\usage{
stop_if_reserved(...)
}
\arguments{
\item{...}{Comma-separated \code{[key]=[value]} pairs}
}
\value{
\code{logical}, indicating if any of the keys was reserved (invisibly).
}
\description{
Utility function checking for reserved names.
}
\section{Details}{

This is a utility function that checks if the keys of the key-value pairs
\code{...} contain reserved words. The reserved words are

\code{.__defaults}, \code{.__reserved}.

If reserved words are encountered in the input an error thrown.
The package vignette has examples of its use:

   \code{vignette('settings',package='options')}
}
\seealso{
\code{\link{is_setting}}
}

