% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.settings.R
\name{update.settings}
\alias{update.settings}
\title{Update Settings}
\usage{
\method{update}{settings}(settings, force = FALSE)
}
\arguments{
\item{settings}{settings file}

\item{force}{Logical: update even if settings have previously been updated?.}
}
\value{
will return the updated settings values
}
\description{
Updates a pecan.xml file to match new layout. This will take care of the
conversion to the latest pecan.xml file.
}
\author{
Rob Kooper
}
