% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.settings.R
\name{write.settings}
\alias{write.settings}
\title{Fix Deprecated Settings}
\usage{
write.settings(settings, outputfile = "pecan.CHECKED.xml")
}
\arguments{
\item{settings}{settings list}
}
\value{
updated settings list
}
\description{
Takes in a settings object, performs a series of checks, fixes & updates settings and produces pecan.CHECKED.xml
}
\author{
Ryan Kelly

Betsy Cowdery
}

