% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.settings.R
\name{addSecrets}
\alias{addSecrets}
\title{Add Users secrets}
\usage{
addSecrets(settings)
}
\arguments{
\item{settings}{settings file}
}
\value{
will return the updated settings values
}
\description{
Add secret information from ~/.pecan.xml
}
\details{
Copies certains sections from ~/.pecan.xml to the settings. This allows
a user to have their own unique parameters also when sharing the
pecan.xml file we don't expose these secrets.
Currently this will copy the database and browndog sections
}
\author{
Rob Kooper
}

