% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.all.settings.R
\name{check.inputs}
\alias{check.inputs}
\title{check to see if inputs are specified - this should be part of the model code}
\usage{
check.inputs(settings)
}
\description{
check to see if inputs are specified - this should be part of the model code
}

