% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.all.settings.R
\name{check.inputs}
\alias{check.inputs}
\title{Check Inputs}
\usage{
check.inputs(settings)
}
\arguments{
\item{settings}{settings file}
}
\description{
check to see if inputs are specified - this should be part of the model code
}
