% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
\name{defaults}
\alias{defaults}
\title{Request default option values}
\usage{
defaults(options)
}
\arguments{
\item{options}{An option manager, as returned by \code{\link{options_manager}} or \code{\link{clone_and_merge}}}
}
\value{
A \code{list}.
}
\description{
Request default option values
}
\seealso{
\code{\link{reset}}
}

