% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSettings.R
\name{expandMultiSettings}
\alias{expandMultiSettings}
\title{generic function for expanding multi-settings.}
\usage{
expandMultiSettings(x)
}
\arguments{
\item{x}{object to be expanded.}
}
\description{
generic function for expanding multi-settings.
}
