% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix.deprecated.settings.R
\name{fix.deprecated.settings}
\alias{fix.deprecated.settings}
\title{Fix Deprecated Settings}
\usage{
fix.deprecated.settings(settings, force = FALSE)
}
\arguments{
\item{settings}{settings list}
}
\value{
updated settings list
}
\description{
Checks for and attempts to fix deprecated settings structure
}
\author{
Ryan Kelly
}
