% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_args.R
\name{get_args}
\alias{get_args}
\title{Get Args}
\usage{
get_args()
}
\value{
list generated by \link[optparse]{parse_args}; see there for details.
}
\description{
Used in web/workflow.R to parse command line arguments.
See also https://github.com/PecanProject/pecan/pull/2626.
}
\examples{
## Example using a script called from your bash terminal:
## If workflow.R contains `get_args()` and settings are in foo.xml, run as
## ./workflow.R -s settings.xml
}
