% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMultisiteMultiSettings.r
\name{getRunSettings}
\alias{getRunSettings}
\title{Build run MultiSettings for a list of site id's}
\usage{
getRunSettings(templateSettings, siteId)
}
\arguments{
\item{templateSettings}{}

\item{siteId}{}
}
\description{
Build run MultiSettings for a list of site id's
}
