% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listToXml.R
\name{listToXml.default}
\alias{listToXml.default}
\title{List to XML}
\usage{
\method{listToXml}{default}(x, ...)
}
\arguments{
\item{x}{object to be converted.
Despite the function name, need not actually be a list}

\item{...}{further arguments.
Used to set the element name of the created XML object,
which is taken from an argument named \code{tag} if present,
or otherwise from the first element of \code{...}}
}
\value{
xmlNode
}
\description{
Convert List to XML
}
\details{
Can convert list or other object to an xml object using xmlNode
}
\author{
David LeBauer, Carl Davidson, Rob Kooper
}
