% Generated by roxygen2 (4.0.2): do not edit by hand
\name{register_options}
\alias{register_options}
\title{Register an option set}
\usage{
register_options(name, defaults = list())
}
\arguments{
\item{name}{name of option set}

\item{defaults}{list of default options}
}
\value{
A \code{function} that can be used as a custom options manager. It has the
following arguments.
\tabular{ll}{
\code{...} \tab Comma separated list of option names (\code{character}) to retrieve options or \code{[name]=[value]} pairs to set options.\cr
\code{.where} \tab (\code{environment}; \code{NULL}) Where to set options (optional, default is global) \code{environment} that was obtained using the \code{clone} or \code{ref} option. \cr
 \code{.simplify}\tab (\code{logical}; \code{TRUE}) Attempt to simplify list of retrieved options?\cr
 \code{.reset} \tab (\code{logical}; \code{FALSE}) Reset options to default values\cr
 \code{.clone} \tab (\code{logical}; \code{FALSE}) Create a copy of the options (returns newly created environment)\cr
 \code{.ref} \tab  (\code{logical}; \code{FALSE}) Return a reference to the option environment.
}
}
\description{
The function \code{register_options} sets up a new options register. It returns
a function that can be uset to set, get, or reset options; or to return a reference to
the options register or to clone the options reference alltogether.
}
\details{
The following terms are arguments of the function that is returned and cannot be used as
option names:

\code{.where} \code{.simplify} \code{.reset} \code{.clone} \code{.ref}
}
\examples{
# create an options register
my_options <- register_options('myopt',defaults=list(foo=1,bar=2,baz='bob'))

### Retrieving global options
my_options() # retrieve the full options list.
my_options('baz')
my_options('foo')

# When multiple options are retrieved, the result is simplified unless specified otherwise:
my_options('foo','baz')
my_options('foo','baz',simplify=FALSE)

### Setting global options
my_options(foo=3,baz='pete')
my_options()

### Create a clone of global options and locally change settings
op <- my_options(clone=TRUE)
## Retrieving local options
my_options(where=op)
my_options('foo','baz',where=op)

## Setting local options
my_options(foo=12,where=op)
my_options('foo',where=op)
# check that global options are unaltered
my_options('foo')

### Create a reference to (global) options
op_ref <- my_options(ref=TRUE)
my_options(foo=0,where=op_ref)
my_options('foo',where=op_ref) #warning: this changes global settings!
my_options('foo')
}

