% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reset}
\alias{reset}
\title{Reset options to default values}
\usage{
reset(options)
}
\arguments{
\item{options}{An option manager, as returned by \code{\link{options_manager}} or \code{\link{clone_and_merge}}}
}
\value{
The list of reset options, invisibly.
}
\description{
Reset options to default values
}
\seealso{
\code{\link{defaults}}
}

