% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setEnsemblePaths.R
\name{setEnsemblePaths}
\alias{setEnsemblePaths}
\title{Set paths to input file ensembles using a consistent pattern across sites}
\usage{
setEnsemblePaths(
  settings,
  n_reps,
  input_type = "met",
  path_template = "./{id}/{n}.nc",
  ...
)
}
\arguments{
\item{settings}{a PEcAn MultiSettings object}

\item{n_reps}{number of replicates to insert for each path.}

\item{input_type}{subsection of the \code{inputs} settings to be edited}

\item{path_template}{format for the paths to be inserted, as a \code{glue} string.}

\item{...}{additional variables to be interpolated into \code{path_template}}
}
\value{
updated settings object
}
\description{
Propagates a filename pattern into the relevant paths of a multi-site
settings. For example if your files are named like "mymet/siteA/scenario1.nc"
up through "mymet/siteZ/scenario50.nc",
\code{setEnsemblePaths(settings, n_reps = 50, "mymet/{id}/scenario{n}.nc")} will
add them all to your settings in one shot.
}
\details{
Operates on one input section (met, poolinitcond, etc) at a time because
it's common to have different path conventions for met vs IC.

The path template should be a string recognized by \code{glue::glue()},
with curly braces wrapping any expressions to be interpolated.
Variables in these expressions are resolved in this order:
\enumerate{
\item Variables defined in each site's \code{run$site} settings block
\item Variables passed as named arguments in \code{...} (same value at every site)
\item \code{{n}} is replaced with the value of \code{1:n_reps}
\item \code{{id}} is replaced with the siteid of each site.
}

If the same name is defined in both run$site and \code{...}, uses the latter.
All variables must have length either 1 or \code{n_reps}.

If \verb{inputs$<input_type>} does not exist, it will be created with a \code{path}
element that matches the requested pattern. If it does exist, any existing
\code{path} element will be overwritten.
}
\examples{
s <- as.Settings(list(
  run = list(
    start.date = "TBD",
    site = list(),
    inputs = list(
      met = list(),
      poolinitcond = list()
    )
  )
))
m <- createMultiSiteSettings(s, c("a1", "b2"))
m1 <- setEnsemblePaths(m, 2)
m1$run$site.a1$inputs
m2 <- setEnsemblePaths(
  m, 2, "poolinitcond",
  icdir = "some/long/path",
  path_template = "{icdir}/{id}/{n}.nc"
)
m2$run$site.a1$inputs

# using site-specific values in the path template:
m_bysite <- createMultiSiteSettings(
  templateSettings = s,
  siteIds = data.frame(id = c("a1", "b2"), grid_cell = c("35N81W", "42N80W"))
)
m2_bysite <- setEnsemblePaths(
  settings = m_bysite,
  n_reps = 2,
  input_type = "met",
  icdir = "some/long/path",
  GCM = c("CESM2", "MIROC6"),
  path_template = "{icdir}/{grid_cell}/{GCM}.nc"
)
m2_bysite$run$site.a1$inputs

}
