% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiSettings.R
\name{settingNames}
\alias{settingNames}
\title{function that can retrieve or update the names of multi-settings.}
\usage{
settingNames(multiSettings, settingNames)
}
\arguments{
\item{multiSettings}{object for which to retrieve or set the names.}

\item{settingNames}{names to be set for the multi-settings object.}
}
\description{
function that can retrieve or update the names of multi-settings.
}
