% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pft_site_linker.R
\name{site.pft.linkage}
\alias{site.pft.linkage}
\title{site.pft.linkage}
\usage{
site.pft.linkage(pecan.xml.address = "", site.pft.links)
}
\arguments{
\item{pecan.xml.address}{character path to the pecan xml file.}

\item{site.pft.links}{dataframe. Your look up table should have two columns of site and pft with site ids under site column and pft names under pft column.}
}
\value{
NONE
}
\description{
This function creates the required tags inside pecan.xml to link sites with pfts given a look up table. If the required tags are already defined in the pecan xml then they will be updated.
}
\examples{
\dontrun{
#setting up the Look up tables
site.pft.links <-tribble(
 ~site, ~pft,
 "1000025731", "temperate.broadleaf.deciduous1",
 "1000025731", "temperate.broadleaf.deciduous11",
 "1000000048", "temperate.broadleaf.deciduous2",
 "772", "temperate.broadleaf.deciduous3",
 "763", "temperate.broadleaf.deciduous4"
)
 
# sending a single setting xml to the function
site.pft.linkage('pecan.xml',site.pft.links)
# sending a multi- setting xml file to the function
site.pft.linkage('pecan.SDA.4site.xml',site.pft.links)
}
}
