% Generated by roxygen2 (4.0.2): do not edit by hand
\name{update.settings}
\alias{update.settings}
\title{Update Settings}
\usage{
\method{update}{settings}(settings)
}
\arguments{
\item{settings}{settings file}
}
\value{
will return the updated settings values
}
\description{
Updates a pecan.xml file to match new layout. This will take care of the
conversion to the latest pecan.xml file.
}
\author{
Rob Kooper
}

