% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.settings.R
\name{update.settings}
\alias{update.settings}
\title{Update Settings}
\usage{
\method{update}{settings}(settings)
}
\arguments{
\item{settings}{settings file}
}
\value{
will return the updated settings values
}
\description{
Updates a pecan.xml file to match new layout. This will take care of the
conversion to the latest pecan.xml file.
}
\author{
Rob Kooper
}

