% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.settings.R
\name{write.settings}
\alias{write.settings}
\title{Write a settings object as XML}
\usage{
write.settings(settings, outputfile, outputdir = settings$outdir)
}
\arguments{
\item{settings}{settings list}

\item{outputfile}{the file name to write to}

\item{outputdir}{the directory to write to}
}
\value{
path to written XML file,
or settings as an XML string if outputfile is NULL
}
\description{
If \code{outputdir} is NULL, the whole path is taken from \code{outputfile}.
If both \code{outputfile} and \code{outputdir} are NULL, no file is written
and it returns the formatted XML as character.
This can be handy for debugging.
}
\author{
Ryan Kelly

Betsy Cowdery
}
