% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.settings.R
\name{write.settings}
\alias{write.settings}
\title{Write settings}
\usage{
write.settings(settings, outputfile, outputdir = settings$outdir)
}
\arguments{
\item{settings}{settings list}

\item{outputfile}{the file name to write to}

\item{outputdir}{the directory to write to}
}
\description{
Takes in a settings object, performs a series of checks, fixes & updates settings and produces pecan.CHECKED.xml
}
\author{
Ryan Kelly

Betsy Cowdery
}
